Add basic product creation
